+++ bank-feeds/RELEASES.md
@@ -119,3 +119,13 @@ Based on:
 - [csharp v2.1.3] bank-feeds
 ### Releases
 - [NuGet v2.1.3] https://www.nuget.org/packages/Codat.BankFeeds/2.1.3 - bank-feeds
+
+## 2023-09-29 00:12:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.2.0] bank-feeds
+### Releases
+- [NuGet v2.2.0] https://www.nuget.org/packages/Codat.BankFeeds/2.2.0 - bank-feeds